# questions/internal/rest/api.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "questionspb/api.proto",
    "version": "version not set"
  },
  "tags": [
    {
      "name": "QuestionsService"
    }
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/api/questions": {
      "post": {
        "summary": "Create a new question",
        "operationId": "createQuestion",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/questionspbCreateQuestionResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/questionspbCreateQuestionRequest"
            }
          }
        ],
        "tags": [
          "Question"
        ]
      }
    },
    "/api/questions/{id}": {
      "get": {
        "summary": "Get a question by id",
        "operationId": "getQuestion",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/questionspbGetQuestionResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "tags": [
          "Question"
        ]
      }
    }
  },
  "definitions": {
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "questionspbCreateQuestionRequest": {
      "type": "object",
      "properties": {
        "body": {
          "type": "string"
        }
      }
    },
    "questionspbCreateQuestionResponse": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "body": {
          "type": "string"
        }
      }
    },
    "questionspbGetQuestionResponse": {
      "type": "object",
      "properties": {
        "question": {
          "$ref": "#/definitions/questionspbQuestion"
        }
      }
    },
    "questionspbQuestion": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "body": {
          "type": "string"
        }
      }
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    }
  }
}